+++ src/main/java/com/example/owetracker/service/ExpenseService.java
@@ -2,6 +2,7 @@ package com.example.owetracker.service;
 
 import com.example.owetracker.model.Expense;
 import com.example.owetracker.model.ExpenseUser;
+import com.example.owetracker.model.ExpensesView;
 import com.example.owetracker.repository.ExpenseRepository;
 import com.example.owetracker.repository.ExpenseUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +10,9 @@ import org.springframework.stereotype.Service;
 import com.example.owetracker.model.User;
 import com.example.owetracker.repository.UserRepository;
 import org.springframework.transaction.annotation.Transactional;
+import java.util.stream.Collectors;
+import java.util.ArrayList;
+import java.util.Comparator;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
@@ -131,11 +135,11 @@ public class ExpenseService {
     }
 
 
-    public List<Expense> getExpensesByGroupId(Long groupId) {
+    public List<Expense> getExpensesByGroupId(Integer groupId) {
         return expenseRepository.findByGroupId(groupId);
     }
 
-    public List<Expense> getExpensesPaidByUser(Long userId) {
+    public List<Expense> getExpensesPaidByUser(Integer userId) {
         return expenseRepository.findByPaidById(userId);
     }
